Fix: typos

+++ src/gl/vgpu/shaderconv.h
@@ -10,7 +10,7 @@
 char * ConvertShaderConditionally(struct shader_s * shader_source, int second_pass);
 char * ConvertShaderVgpu(struct shader_s * shader_source, int second_pass);
 char * ConvertShaderMinimal(char * input, int is_fragment);
-char * ConvertShaderMinimalBackport(char * input, int * length, int is_fragment, int destructive);
+char * ConvertShaderMinimalBackport(char * input, size_t * length, int is_fragment, int destructive);
 char * BackportAttributes(char * source, int *sourceLength);
 
 char * GLSLHeader(char* source);
